refactor: simplify internal/ directory structure
Consolidate internal packages from 4 to 2 for better cohesion:

- Remove internal/testutil package: Move SQL loading functions directly
  into dbtest/test/helpers.go where they're only used
- Merge internal/spanner into internal/validator: Embed Spanner client
  functionality as unexported methods since validator is the only consumer
- Add new NewValidator() constructor that creates Spanner client internally
- Maintain backward compatibility with existing New() constructor
- Update all imports and type references

Benefits:
- Reduced package complexity (4 → 2 packages)
- Better cohesion between related functionality
- Cleaner dependency structure
- All tests pass, application builds successfully
